backend/requirements.txt
# 🔹 기본 패키지
fastapi
uvicorn
python-dotenv

# 🔹 인증 관련
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]



# 🔹 NLP/AI 관련
#transformers==4.36.2
#sentence-transformers==2.2.2

# 🔹 Langchain 관련 패키지
langchain
langchain-community
langchain-openai
langchain-weaviate
langchain-huggingface
openai

# 🔹 기타
openai
weaviate-client
psycopg2-binary

